refactor: fix policy msg typo and abstract commons test functions
